Move funcs to util.js
