+++ src/util.js
@@ -96,6 +96,40 @@ async function userHasItem(id, item, db) {
     return itemResult != undefined
 }
 
+const handlePlaceholders = (str, objs = {}) => {
+    if (objs.source.args) str = replaceArgPlaceholders(str, objs.source.args)
+    if (objs.source) str = replaceEventPlaceholders(str, objs.source)
+    if (objs.opts) str = replaceOptsPlaceholders(str, objs.opts)
+    return str
+}
+
+const replaceArgPlaceholders = (str, args) => {
+    for (let i = 0; i < args.length; i++) {
+        const re = new RegExp(escapeRegexSpecialChars('$arg' + i), 'g')
+        str = str.replace(re, args[i])
+    }
+    return str
+}
+const replaceEventPlaceholders = (str, source) => {
+    var keys = Object.keys(source)
+    keys.forEach(key => {
+        if (key == 'args') return
+        const re = new RegExp(escapeRegexSpecialChars('$:' + key), 'g')
+        str = str.replace(re, source[key])
+    })
+    return str
+}
+
+const replaceOptsPlaceholders = (str, opts) => {
+    var keys = Object.keys(opts)
+    keys.forEach(key => {
+        if (safeToString(opts[key]) == '[Object object]') return
+        const re = new RegExp(escapeRegexSpecialChars('$_' + key), 'g')
+        str = str.replace(re, safeToString(opts[key]))
+    })
+    return str
+}
+
 const replacePlaceholders = (str, placeholders) => {
     Object.keys(placeholders).forEach(p => {
         const re = new RegExp(escapeRegexSpecialChars(p), 'g')
@@ -110,6 +144,7 @@ module.exports = {
     timeObjToMs,
     checkCooldown,
 
+    handlePlaceholders,
     replacePlaceholders,
     escapeMarkdown,
     attachmentType,
